BJS2-63054 правки по комментам

+++ src/main/java/school/faang/helpanimal/Donation.java
@@ -1,9 +1,10 @@
 package school.faang.helpanimal;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record Donation(UUID id, int amount) {
-    public Donation(int amount) {
+public record Donation(UUID id, BigDecimal amount) {
+    public Donation(BigDecimal amount) {
         this(UUID.randomUUID(), amount);
     }
 }
